+++ node/src/Transaction.ts
@@ -130,6 +130,7 @@ import {
     createRPop,
     createRPush,
     createRPushX,
+    createRandomKey,
     createRename,
     createRenameNX,
     createSAdd,
@@ -2503,6 +2504,17 @@ export class BaseTransaction<T extends BaseTransaction<T>> {
      */
     public geohash(key: string, members: string[]): T {
         return this.addAndReturn(createGeoHash(key, members));
+    }
+
+    /**
+     * Returns a random existing key name from the currently selected database.
+     *
+     * See https://valkey.io/commands/randomkey/ for more details.
+     *
+     * Command Response - A random existing key name from the currently selected database.
+     */
+    public randomKey(): T {
+        return this.addAndReturn(createRandomKey());
     }
 }